+++ c++/include/DynamicObstacleAvoidance/Obstacle/Obstacle.hpp
@@ -19,11 +19,12 @@ private:
 
 	Eigen::Vector3f linear_velocity;
 	Eigen::Vector3f angular_velocity;
-	double safety_margin;
+	
+	float safety_margin;
 
 public:
-	explicit Obstacle(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, Eigen::Vector3f& reference_position, double safety_margin=0);
-	explicit Obstacle(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, double safety_margin=0);
+	explicit Obstacle(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, Eigen::Vector3f& reference_position, float safety_margin=0);
+	explicit Obstacle(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, float safety_margin=0);
 	~Obstacle();
 
 	inline const Eigen::Vector3f get_center_position() const 
@@ -51,14 +52,14 @@ public:
 		return this->angular_velocity;
 	}
 
-	inline double get_safety_margin() const
+	inline float get_safety_margin() const
 	{ 
 		return this->safety_margin;
 	}
 
 	virtual Eigen::Vector3f compute_normal_to_external_point(const Eigen::Vector3f& external_point) const;
 	
-	virtual double compute_distance_to_external_point(const Eigen::Vector3f& external_point) const;
+	virtual float compute_distance_to_external_point(const Eigen::Vector3f& external_point) const;
 };
 
 #endif
