+++ c++/src/Obstacle/Ellipsoid2D.cpp
@@ -1,13 +1,13 @@
 #include "DynamicObstacleAvoidance/Obstacle/Ellipsoid2D.hpp"
 
-Ellipsoid2D::Ellipsoid2D(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, Eigen::Vector3f& reference_position, double safety_margin):
+Ellipsoid2D::Ellipsoid2D(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, Eigen::Vector3f& reference_position, float safety_margin):
 Obstacle(center_position, center_orientation, reference_position, safety_margin)
 {
 	this->axis_lengths << 1, 1;
 	this->curvature_factor << 1, 1;
 }
 
-Ellipsoid2D::Ellipsoid2D(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, double safety_margin):
+Ellipsoid2D::Ellipsoid2D(Eigen::Vector3f& center_position, Eigen::Vector4f& center_orientation, float safety_margin):
 Ellipsoid2D(center_position, center_orientation, center_position, safety_margin)
 {}
 
@@ -24,7 +24,7 @@ Eigen::Vector3f Ellipsoid2D::compute_normal_to_external_point(const Eigen::Vecto
 	return normal_vector;
 }
 
-double Ellipsoid2D::compute_distance_to_external_point(const Eigen::Vector3f& external_point) const
+float Ellipsoid2D::compute_distance_to_external_point(const Eigen::Vector3f& external_point) const
 {
 	Eigen::Array3f point_in_frame = external_point - this->get_center_position();
 	Eigen::Array2f tmp_values = point_in_frame.head(2) / this->axis_lengths;
